risk_engine/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
supabase==2.0.2
numpy==1.24.3
pandas==2.0.3
scipy==1.11.1
matplotlib==3.7.2
seaborn==0.12.2
yfinance==0.2.28
requests==2.31.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
asyncio==3.4.3
aiofiles==23.2.1
Pillow==10.0.0
plotly==5.17.0
kaleido==0.2.1
tabulate==0.9.0
PyJWT==2.8.0
